rename block_to_json.rs to json_block.rs
