+++ query/src/servers/http/v1/mod.rs
@@ -12,20 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod block_to_json;
 mod http_query_handlers;
+pub mod json_block;
 mod load;
 mod query;
 mod stage;
 mod statement;
 
-pub(crate) use block_to_json::JsonBlock;
 pub use http_query_handlers::make_final_uri;
 pub use http_query_handlers::make_page_uri;
 pub use http_query_handlers::make_state_uri;
 pub use http_query_handlers::query_route;
 pub use http_query_handlers::QueryResponse;
 pub use http_query_handlers::QueryStats;
+pub(crate) use json_block::JsonBlock;
 pub use load::streaming_load;
 pub use load::LoadResponse;
 pub use query::ExecuteStateKind;
