+++ query/tests/it/servers/http/mod.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod block_to_json;
 mod clickhouse_handler;
 mod formats;
 mod http_query_handlers;
+mod json_block;
 mod statement;
